feat: fetch bike station data on utils
